Substitute AzureOAuth2ServiceTest to OAuth2AuthenticatorTest
