+++ tests/AdminAuthServiceTest.php
@@ -1,9 +1,13 @@
 <?php
 declare(strict_types=1);
 
-namespace Ridibooks\Cms\Service;
+namespace Ridibooks\Cms\Tests;
 
 use PHPUnit\Framework\TestCase;
+use Ridibooks\Cms\Service\AdminAuthService;
+use Ridibooks\Cms\Service\AdminMenuService;
+use Ridibooks\Cms\Service\AdminTagService;
+use Ridibooks\Cms\Service\AdminUserService;
 use Ridibooks\Cms\Thrift\AdminUser\AdminUser;
 use Ridibooks\Cms\Thrift\Errors\UnauthorizedException;
