+++ tests/AdminUserServiceTest.php
@@ -1,9 +1,10 @@
 <?php
 declare(strict_types=1);
 
-namespace Ridibooks\Cms\Service;
+namespace Ridibooks\Cms\Tests;
 
 use PHPUnit\Framework\TestCase;
+use Ridibooks\Cms\Service\AdminUserService;
 
 // TODO(devgrapher): temparary test cases. Model classes are statically bound inside. Hard to inject mocks.
 class AdminUserServiceTest extends TestCase
